+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Implemented darkened VUs when muted/faded in MixMasters and AuxSpanders
 - Added initialization of individual track settings in MixMasters and EqMaster
 - Made filter mappable (click filter lights when mapping)
+- Added option in EQMaster menu to hide EQ curves when bypassed, for when EQ is to be used as a spectrum analyzer only
 
 
 ### 1.1.5 (2020-01-29)